Squashed commit of the following:

commit a5e68a2473906fa16c21af0fca541504ce54bfc3
Author: Alexander Lavrukov <lavrukov@yandex-team.ru>
Date:   Mon Oct 19 22:02:24 2020 +0300

    CLOUD-56633: improve sg sweeper

+++ yandex/resource_yandex_vpc_security_group_test.go
@@ -58,6 +58,17 @@ func sweepVPCSecurityGroupOnce(conf *Config, id string) error {
 	ctx, cancel := conf.ContextWithTimeout(yandexVPCNetworkDefaultTimeout)
 	defer cancel()
 
+	sg, err := conf.sdk.VPC().SecurityGroup().Get(ctx, &vpc.GetSecurityGroupRequest{
+		SecurityGroupId: id,
+	})
+	if err != nil {
+		return err
+	}
+
+	if sg.DefaultForNetwork {
+		return nil
+	}
+
 	op, err := conf.sdk.VPC().SecurityGroup().Delete(ctx, &vpc.DeleteSecurityGroupRequest{
 		SecurityGroupId: id,
 	})
